chore(webapp): rename trade dialog to be close specific
