+++ webapp/frontend/lib/trade/close_position_confirmation_dialog.dart
