+++ webapp/frontend/lib/trade/position_table.dart
@@ -4,11 +4,11 @@ import 'package:get_10101/common/direction.dart';
 import 'package:get_10101/common/model.dart';
 import 'package:get_10101/settings/channel_change_notifier.dart';
 import 'package:get_10101/settings/dlc_channel.dart';
+import 'package:get_10101/trade/close_position_confirmation_dialog.dart';
 import 'package:get_10101/trade/position_change_notifier.dart';
 import 'package:get_10101/trade/position_service.dart';
 import 'package:get_10101/trade/quote_change_notifier.dart';
 import 'package:get_10101/trade/quote_service.dart';
-import 'package:get_10101/trade/trade_confirmation_dialog.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import 'package:collection/collection.dart';
